index.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>NFC test</title>
    <script src="./resources/testharness.js"></script>
    <script src="./resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
        //test(function() {
        //    assert_true(document.implementation.hasFeature())
        //},"hasFeature() with no arguments")
        //test();

        //async_test(function(t) {
        //    document.addEventListener("DOMContentLoaded", function() {
        //        t.step(function() {
        //            assert_true(t.bubbles, "bubbles should be true");
        //        });
        //        t.done();
        //    });
        //},"DOMContentLoaded");

        function foo() {
            return Promise.resolve("foo");
        }
        promise_test(function() {
            return foo()
                .then(function(result) {
                    assert_equals(result, "foo", "foo should return 'foo'");
                });
        },"simple example");
        function bar() {
            return Promise.reject(new TypeError());
        }
        promise_test(function(t) {
            return promise_reject(t, new TypeError(), bar());
        }, "another example")
    </script>
  </body>
</html>
